refactor(login): co-locate login component with auth provider
