+++ src/components/login.tsx
@@ -2,7 +2,7 @@ import { Button, Card, Help, InputField, LogoIcon } from '@dhis2/ui'
 import { useCallback, useState } from 'react'
 import { Link } from 'react-router-dom'
 import styles from './login.module.css'
-import { useAuth } from '../../hooks'
+import { useAuth } from '../hooks'
 
 export const Login = () => {
     const { login, isAuthenticating, authenticationErrorMessage } = useAuth()
